+++ zio-examples/build.sbt
@@ -1,6 +1,8 @@
 lazy val threadlocalBridge = RootProject(file("threadlocal-bridge"))
 
+lazy val reloadableServices = RootProject(file("reloadable-services"))
+
 lazy val root = project
   .in(file("."))
   .settings(publish / skip := true)
-  .aggregate(threadlocalBridge)
+  .aggregate(threadlocalBridge, reloadableServices)
