+++ zio-examples/reloadable-services/build.sbt
@@ -1,0 +1,6 @@
+scalaVersion := "2.13.18"
+publish / skip := true
+
+libraryDependencies ++= Seq(
+  "dev.zio" %% "zio" % "2.1.25"
+)
